rename ColumnFingerprintInfo folder too
