+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnFingerprintInfo/DateTimeFingerprint.tsx
