+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnFingerprintInfo/GlobalFingerprint.styled.tsx
