+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnFingerprintInfo/GlobalFingerprint.tsx
