+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnFingerprintInfo/GlobalFingerprint.unit.spec.tsx
