+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnFingerprintInfo/QueryColumnFingerprintInfo.tsx
