+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnFingerprintInfo/QueryColumnFingerprintInfo.unit.spec.tsx
