+++ frontend/src/metabase/common/components/MetadataInfo/QueryColumnFingerprintInfo/index.ts
